Final_project/requirements.txt
gensim
kafka-python
lightgbm
numpy
pandas

